GeoJsonService

+++ src/gView.GeoJsonService/DTOs/GetMapResponse.cs
@@ -5,11 +5,11 @@ public class GetMapResponse
     public string Type { get; set; } = "GetMapResponse";
     public string? ImageUrl { get; set; }
     public string? ImageBase64 { get; set; }
-    public float[] BBox { get; set; } = Array.Empty<float>();
+    public double[] BBox { get; set; } = Array.Empty<double>();
     public int Width { get; set; }
     public int Height { get; set; }
-    public float ScaleDenominator { get; set; }
-    public float? Rotation { get; set; }
+    public double ScaleDenominator { get; set; }
+    public double? Rotation { get; set; }
     public string Format { get; set; } = string.Empty;
 }
